Join all transformations in an integration “preset”
Naming taking from babel, although not exactly equivalent.
